backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pdfminer.six==20240706
docx2txt==0.8
openai==1.55.3